requirements.txt
# RDKit and core dependencies
# Note: RDKit is best installed via conda: conda install -c conda-forge rdkit
# If using pip: rdkit-pypi (may have limited functionality)
rdkit-pypi>=2023.9.1

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
pillow>=10.0.0

# Development tools
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
